Add support for default tasks
If only a single argument is provided, try to do something sensible. The
argument is assumed to be the source directory.  That directory is
searched for a tasks.yml file to use.  If the tasks file cannot be
found, a default task is created which will symlink all files/folders
found in the source directory to the destination.